making updates so that health tests run faster
